refactor(PreparedQuantity): update onPreparedQuantityChange type and apply className dynamically refactor: migrate cn utility imports from legacy to shared module
