+++ src/sections/recipe/components/RecipeEditView.tsx
@@ -27,7 +27,7 @@ import { PreparedQuantity } from '~/sections/common/components/PreparedQuantity'
 import { useFloatField } from '~/sections/common/hooks/useField'
 import { FloatInput } from '~/sections/common/components/FloatInput'
 import { RecipeEditor } from '~/legacy/utils/data/recipeEditor'
-import { cn } from '~/legacy/utils/cn'
+import { cn } from '~/shared/cn'
 import { type JSXElement, type Accessor, type Setter } from 'solid-js'
 
 export type RecipeEditViewProps = {
